upgraded gradle and renamed duplicate namespaces
