+++ build.gradle
@@ -1,36 +1,45 @@
 plugins {
-    id 'com.jfrog.bintray' version '1.8.4'
-    id 'com.github.ben-manes.versions' version '0.25.0'
+    id 'maven-publish'
+    id 'java-library'
+    id 'java'
+    id 'groovy'
 }
 
-apply plugin: 'java'
-apply plugin: 'groovy'
-apply plugin: 'maven'
+group = 'no.fintlabs'
 
 sourceCompatibility = 1.8
 targetCompatibility = 1.8
 
 repositories {
-    maven {
-        url "http://dl.bintray.com/fint/maven"
-    }
     mavenLocal()
+    maven {
+        url "https://repo.fintlabs.no/releases"
+    }
     mavenCentral()
 }
 
-apply from: 'https://raw.githubusercontent.com/FINTlibs/fint-buildscripts/v1.7.0/dependencies.gradle'
+publishing {
+    publications {
+        mavenJava(MavenPublication) {
+            from components.java
+        }
+    }
+    repositories {
+        mavenLocal()
+    }
+}
+
+configurations {
+    compileOnly {
+        extendsFrom annotationProcessor
+    }
+}
+
 dependencies {
-    compile('no.fint:fint-relation-model:1.1.5')
-    compile('no.fint:fint-model-core:0.1.0')
-    compile('no.fint:fint-model-resource:0.3.3')
-    compile('com.fasterxml.jackson.core:jackson-annotations:2.8.0')
-    compileOnly("org.projectlombok:lombok:${lombokVersion}")
-}
-
-task wrapper(type: Wrapper) {
-    gradleVersion = gradleVersion
-}
-
-if (project.hasProperty('bintrayUser') && project.hasProperty('bintrayKey')) {
-    apply from: 'https://raw.githubusercontent.com/FINTlibs/fint-buildscripts/v1.7.0/bintray.gradle'
+    implementation('no.fint:fint-relation-model:1.1.6')
+    implementation('no.fint:fint-model-core:0.2.1')
+    implementation('no.fint:fint-model-resource:0.4.1')
+    implementation('com.fasterxml.jackson.core:jackson-annotations:2.8.0')
+    compileOnly("org.projectlombok:lombok:1.18.24")
+    annotationProcessor 'org.projectlombok:lombok:1.18.24'
 }
