+++ gradle.properties
@@ -1,1 +1,1 @@
-version=0.1.1
+version=0-SNAPSHOT
