+++ src/main/java/no/fint/model/pwfa/Dog.java
@@ -1,20 +1,13 @@
 // Built from tag v0.1.1
 
-package no.fint.model.pwfa.pwfa;
+package no.fint.model.pwfa;
 
-import lombok.Data;
-import lombok.NoArgsConstructor;
-import lombok.EqualsAndHashCode;
-import lombok.ToString;
-import lombok.NonNull;
-import java.util.List;
+import lombok.*;
 import no.fint.model.FintMainObject;
-import no.fint.model.pwfa.pwfa.Identifikator;
 
 @Data
 @NoArgsConstructor
-@EqualsAndHashCode
-@ToString
+@AllArgsConstructor
 public class Dog implements FintMainObject {
     public enum Relasjonsnavn {
             OWNER
