+++ src/main/java/no/fint/model/resource/pwfa/DogResource.java
@@ -1,6 +1,6 @@
 // Built from tag v0.1.1
 
-package no.fint.model.resource.pwfa.pwfa;
+package no.fint.model.resource.pwfa;
 
 import com.fasterxml.jackson.annotation.JsonIgnore;
 
@@ -18,14 +18,15 @@ import java.util.Map;
 import no.fint.model.FintMainObject;
 import no.fint.model.resource.FintLinks;
 import no.fint.model.resource.Link;
-import no.fint.model.pwfa.pwfa.Identifikator;
+import no.fint.model.pwfa.Identifikator;
 
 @Data
 @NoArgsConstructor
 @EqualsAndHashCode
 @ToString
-public class OwnerResource implements FintMainObject, FintLinks {
+public class DogResource implements FintMainObject, FintLinks {
     // Attributes
+    private String breed;
     @NonNull
     private Identifikator id;
     @NonNull
@@ -36,10 +37,10 @@ public class OwnerResource implements FintMainObject, FintLinks {
     private final Map<String, List<Link>> links = createLinks();
         
     @JsonIgnore
-    public List<Link> getDog() {
-        return getLinks().getOrDefault("dog", Collections.emptyList()); 
+    public List<Link> getOwner() {
+        return getLinks().getOrDefault("owner", Collections.emptyList()); 
     }
-    public void addDog(Link link) {
-        addLink("dog", link);
+    public void addOwner(Link link) {
+        addLink("owner", link);
     }
 }
