+++ src/main/java/no/fint/model/resource/pwfa/OwnerResources.java
@@ -1,6 +1,6 @@
 // Built from tag v0.1.1
 
-package no.fint.model.resource.pwfa.pwfa;
+package no.fint.model.resource.pwfa;
 
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.fasterxml.jackson.core.type.TypeReference;
@@ -12,16 +12,16 @@ import lombok.NoArgsConstructor;
 import no.fint.model.resource.AbstractCollectionResources;
 
 @NoArgsConstructor
-public class DogResources extends AbstractCollectionResources<DogResource> {
+public class OwnerResources extends AbstractCollectionResources<OwnerResource> {
 
-    public DogResources(Collection<DogResource> input) {
+    public OwnerResources(Collection<OwnerResource> input) {
         super(input);
     }
 
     @JsonIgnore
     @Deprecated
     @Override
-    public TypeReference<List<DogResource>> getTypeReference() {
-        return new TypeReference<List<DogResource>>() {};
+    public TypeReference<List<OwnerResource>> getTypeReference() {
+        return new TypeReference<List<OwnerResource>>() {};
     }
 }
